Break application into modules
[__init__.py]
Contains the Flask application.

[db.py]
Move database functionalities to db module.

[views.py]
Move view functions to views module.

[server.py]
Run the application.
