+++ server.py
@@ -1,0 +1,1 @@
+from quiz import app
